Multiple CF fix
