+++ WalletWasabi.Fluent/ViewModels/Wallets/Buy/Messages/HyperlinkViewModel.cs
@@ -1,0 +1,25 @@
+using System.Threading.Tasks;
+using System.Windows.Input;
+using ReactiveUI;
+using WalletWasabi.Fluent.Helpers;
+using WalletWasabi.Helpers;
+
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Buy.Messages;
+
+internal class HyperlinkViewModel
+{
+	public string Text { get; }
+	public string Url { get; }
+
+	public HyperlinkViewModel(string text, string url)
+	{
+		Text = text;
+		Url = url;
+		OpenLinkCommand = ReactiveCommand.Create(() => IoHelpers.OpenBrowserAsync(url));
+		CopyCommand = ReactiveCommand.CreateFromTask(() => ApplicationHelper.SetTextAsync(Url) ?? Task.CompletedTask);
+	}
+
+	public ICommand OpenLinkCommand { get; set; }
+
+	public ICommand CopyCommand { get; set; }
+}
