+++ WalletWasabi/WebClients/ShopWare/Models/ShoppingCart.cs
@@ -3,4 +3,3 @@ namespace WalletWasabi.WebClients.ShopWare.Models;
 public record ShoppingCartResponse(string Token);
 
 public record ShoppingCartItemsResponse (string Token);
-
